requirements.txt
# Core
requests>=2.28.0
dnspython>=2.3.0
colorama>=0.4.6
urllib3>=1.26.0
pyyaml>=6.0

# Fase 2 - Brute-force & Tech Detection
aiohttp>=3.8.4
asyncio>=3.4.3
tqdm>=4.65.0
beautifulsoup4>=4.12.0
whattheversion>=1.0.0

# Fase 3 - Port Scanning & Screenshots
python-nmap>=0.7.1
selenium>=4.9.0
webdriver-manager>=3.8.6
Pillow>=9.5.0

# Fase 4 - Takeover Detection & Reporting
jinja2>=3.1.2
markdown>=3.4.3
python-whois>=0.8.0